Determine whether a filter should be displayed or not (based on whether any content is available for it).
Dirty code just to see how it performs with a real data. On paper, this has horrible performance, with a worst case of O(N*M*P). However, it can likely be optmized (a bit), the result is very cacheable, and I suspect that the real world performance will be much better.